Extract and export config.
